requirements.txt
# Core framework
streamlit>=1.28.0
crewai>=0.22.0
crewai-tools>=0.1.0

# Environment and configuration
python-dotenv>=1.0.0

# AI/ML libraries
openai>=1.0.0

# Google APIs for contacts integration
google-auth>=2.0.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0

# Data processing
pydantic>=2.0.0

# Utility libraries
wikipedia>=1.4.0
schedule>=1.2.0

# Email validation
validate_email>=1.3.0
dnspython>=2.0.0

# Additional utilities
pathlib2>=2.3.0
typing-extensions>=4.0.0